bot/requirements.txt
# Discord Bot
discord.py>=2.3.0
python-dotenv>=1.0.0

# Google Sheets Integration
gspread>=5.12.0
oauth2client>=4.1.3

# Web Scraping
selenium>=4.15.0
webdriver-manager>=4.0.1

# Database
# sqlite3 is built into Python

# Utilities
requests>=2.31.0
